requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
WTForms==3.0.1
Werkzeug==2.3.7
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
flask-caching==2.1.0
flask-cors==4.0.0
cryptography==41.0.7
flasgger==0.9.7.1
python-dateutil==2.8.2
pytz==2023.3
